Remove caml_block and caml_tuple
Summary:
These helpers are convenient, but in typical usage, they use a lot of stack space. The caller typically does something like this:

```
        caml_tuple(&[
            context.source_text,
            usize_to_ocaml(self.offset),
            usize_to_ocaml(self.leading_width),
            usize_to_ocaml(self.width),
            usize_to_ocaml(self.trailing_width),
            to_list(&self.leading, context),
            to_list(&self.trailing, context),
        ])
```

But this means that the caller stack-allocates an array of 7 words (56 bytes), populates it, and hands a pointer and length to caml_tuple, which must copy each word into the actual allocated block. I suspect that we might be paying a large total stack-space cost for deeply mutually recursive serialization functions which stack-allocate arrays in this way.

Instead, we can use `reserve_block` and `caml_set_field` directly. The code is a little uglier, but we use less stack space and avoid the need to copy the fields from the stack into the newly allocated block.

Reviewed By: shiqicao

Differential Revision: D18171912

fbshipit-source-id: d3013a2fdd835b71309fe85b7e4c3559b1f2618e