Avoid crashing hh_client on hh_server restart
Summary:
As reported by Daniel Ohayon, serverless IDE sometimes becomes unresponsive after big rebases.  Since big rebases often trigger hh_server to halt and restart, I investigated the behavior of serverless IDE in scenarios where hh_server stops and starts.

In testing, I was able to demonstrate that hh_client will self-terminate when hh_server relaunches.  When hh_client terminates, serverless IDE is also halted.  The repro is this:

* Launch hh_client and hh_server
* Ensure both are in a stable state and IDE services are functioning
* Halt hh_server using `hh stop`
* Ensure that IDE services continue to function
* Restart hh_server using `hh`

In Daniel's experience, doing this change a few times repeatedly would cause the Nuclide server to disconnect from hh_client and stop sending LSP requests.

Logs demonstrated that the reason for hh_client self-terminating is because it believed there were unsaved changes in the editor, which triggered a call to `exit_fail`.

After this diff, hh_client remains up after hh_server restarts and no `exit_fail` is triggered.

The reason for this behavior, according to Lucian Wischik, is that in the past it was necessary to restart hh_client in order to get a reliable set of edit-file messages because hh_client did not store complete editor state in memory.

Since we now track that state in memory, we can now send that information on startup and we no longer need to terminate hh_client.  I generate that state using OPEN_FILE messages with the full text of the file, since there's no need to try to track individual edits.

Reviewed By: arxanas

Differential Revision: D18205060

fbshipit-source-id: 04675f002ed6d07fa4cf479e24e9563bd73aca34