Make the endo visitor generic over AAST
Summary: When I originally created the visitor - I made it operate over the nast. But in hh_single_compile we go directly from the legacy ast to the [empty] tast. I played around with the object/class notation in ocaml and made the namespace elaborator generic over the aast. I plan to stick this step in front of codegen steps and see what breaks.

Reviewed By: kmeht

Differential Revision: D17865869

fbshipit-source-id: 0b733072081e931500d2301d45ec18a8b879c894