Elaborate typehints
Summary:
We currently perform namespace elaboration all over the place - such as in naming and in codegen. If we are able to extract the namespace elaboration into one pass, we can remove having separate logic in naming and codegen and have it all in one place.
The plan is to
1.) Create an endo visitor that performs just namespace elaboration
2.) Move all of the elaboration from naming into it
3.) Add this pass into the codegen pass and see if codegen produces different bytecode

Elaborates typehints.

Diff concludes all elaboration that happened in naming where the elaborated name appeared in the tree.
We can delete the temporary holdover function `type_name_without_elaboration` since all uses of `type_name` in naming no longer perform elaboration.

Reviewed By: kmeht

Differential Revision: D17847136

fbshipit-source-id: 7a5a1b3b2a4a7a3317620232f5a09f9b49e34166