Remove cruft left over from stream user filter removal.
Summary: ext_stream-user-filters.php declares some functions that are no longer defined (calling them will blow up), so delete it, and delete mentions of those same functions in builtins_stream.hhi. Also delete a test that was trying to test the deleted functions, but was always skipping because it's skipif file tried to call one of them and always blew up.

Reviewed By: ricklavoie

Differential Revision: D18213424

fbshipit-source-id: dfe4acfa0bfdcd0f777183c303ec108b89eb821a