Simplify set_pre_timeout_handler's behavior
Summary:
This change makes set_pre_timeout_handler somewhat equivalent of Javascript's "setTimeout" function. In other words, it allows the userland code to schedule a callback in X seconds. Those callbacks can be used to conditionally perform expensive operations later without blocking the request if the request is shorter. One example is do to expensive stack trace logging right before timeouts.
This simplifies the behavior of the timer and the way we reset it.
This diff also adds a few locations where the timer should have been reset to a specific value.

Differential Revision: D18208541

fbshipit-source-id: c2f0e86b9d3078319d093b097e8fa5011ab8dc7c