Fix and re-enable static string tracking in MixedArrayKeys
Summary:
In MakeUncounted, we can create add static string keys to the array. That causes a bug when static-string tracking is on in MixedArrayKeys, since we don't update that bitset in this function.

To fix the issue, I start off making all keys in the bitset static, and then if we are forced to put an uncounted string in, I record that. This approach doesn't add any branches to the loop, and it also has the nice effect of keeping as few bits set in MixedArrayKeys as possible. I also noticed that I named a function "makeUncounted" when it's actually "makeStatic" - uncounted MixedArrays may have non-static keys, but static MixedArrays may not - so I renamed that function.

With the bug fixed, we can re-enable this feature. (Once on, it's guarded by assertions which should help us catch any future failures early.) The benefit of enabling this feature is that we can check that a MixedArray has all int keys in O(1) time and thus specialize more loops - see the performance numbers in D18194604.

Reviewed By: mofarrell

Differential Revision: D18094964

fbshipit-source-id: 8da11fc990789aeaf2954679ff9b1c40a24af786