Removed trace from direct decl parser.
Summary: In the beginning, when I was getting the skeleton for direct parsing set up, it was useful to be able to log the tree that we returned.  Now that that base is set up (and a lot of nodes end up getting converted into `Node_::Ignore` after the decl they correspond to is generated), it's not really as useful as just manually instrumenting the smart constructors with `println!()` macros, and it's very noisy when all we want is timing data.

Differential Revision: D18144551

fbshipit-source-id: e73b630980e16d65f56fd1a07833d70956910789