Print timing information in hh_single_decl.
Summary: We have a way to compare direct decl parsing against the existing parsing, but we also need some way to compare it to the existing file -> decl pipeline, which is what it will actually be replacing.  `hh_single_decl` already compares these two for correctness, so it seems like a logical place to compare them for speed too.

Differential Revision: D18144549

fbshipit-source-id: ca832b3ce41076682d46bedb38f69a31e6e0bcc5