Replace toC?(Arr|Str|Obj)Ref with asC?(Arr|Str|Obj)Ref
Summary: Now that we no longer have refs, the behavior of all the toC?(Arr|Str|Obj)Ref helpers is identical to that of their respective asC?(Arr|Str|Obj)Ref helpers. Since these are composed of only asserts + casts, 'as' seems to be a better naming convention for them. Switch everything to the 'as' versions and remove the 'to' versions.

Reviewed By: ricklavoie

Differential Revision: D18204093

fbshipit-source-id: dd592974ea5c2505f2cc43bd83c1e9720e598ff5