Simplify arrayRefShuffle.
Summary:
arrayRefShuffle was a bit of a mess:
 * it contained dead code related to handling a set into an array with refs in it
 * it was trying to handle 2 almost unrelated use cases: the jit callers always had false for the setRef template param and always passed a nullptr base (i.e. all they wanted was a decref of oldData if it wasn't the same as newData), and the interp callers always had true for the setRef template param and a valid lval for the base (i.e. in addition to the refcounting they wanted the base to be updated when oldData != newData).

Have the jit callers do the decref themseves: this operation is small enough to not need a helper.

For the interp callers: rename arrayRefShuffle to arraySetUpdateBase and assert that the base contained oldData instead of assuming the base may have been a ref that in turn contained oldData (and checking whether the base (still) contains oldData).

Reviewed By: ricklavoie

Differential Revision: D18210437

fbshipit-source-id: 02ec4f9b799a7f8227dbf4f8e600e54391709de6