Cleanup 3 crappy tests
Summary:
Those tests have other errors (missing visibility specifiers) besides the error that they are supposed to test. This gets confusing, because in tests we only print the first error, so a change to ordering can completely change test output for seemingly no reason.

Fixing the errors by adding the missing modifiers.

`typing_fail_test.php` is completely deleted, because it seems like it belongs in museum, not a test suite. It's very ancient, uses syntax that nobody even remembers and it's unclear what it is testing at this point.

Reviewed By: shiqicao

Differential Revision: D18237188

fbshipit-source-id: b297597c1fb95561f5d71b7fe1d5dc707ea67cc0