Model HH-autoimport of types as 'use type' statements
Summary:
As a result of past work to standardize HH-autoimporting between the typechecker and compiler, it is semantically equivalent to a series of `use {type, function, const, namespace}` statements at the top of every file and/or namespace block (except that they can be overridden by users in code).

This diff literally considers type autoimports to be `use type` statements, which allows us to eliminate special-casing for HH-autoimporting and reuse the existing namespace elaboration machinery.

There is a subtle breaking change here—users may no longer override primitives with `use` statements; e.g. the literal `string` will *always* refer to the string primitive. They can, however, use "String", because ID elaboration is case-sensitive.

Differential Revision: D18188079

fbshipit-source-id: c51fdf35ade48998fa500e507344a83d6bde4912