Remove unnecessary namespace elaboration in Rx
Summary: Now that we are elaborating names earlier in the pipeline in HackC (previous diff in stack), we can begin to remove the places where we are elaborating names in codegen. This diff removes one instance. We will have already elaborated the name by the time we reach this point.

Reviewed By: kmeht, losvald

Differential Revision: D18020994

fbshipit-source-id: 1467f6733bc3111c6bcf961a9455919ca9393c2f