Add namespace elaboration pass before codegen
Summary:
In the diffs below this - we extracted namespace elaboration from naming into its own separate ast pass in `elaborate_namespaces_endo.ml`.

In the following diffs, we want to have codegen go through the same pass and remove the namespace elaboration from the codegen code.

In this diff, we add the namespace elaboration pass in front of the codegen pass without removing any of the namespace elaboration. This is okay because namespace elaboration is designed to be idempotent. The only change should be that we perform namespace elaboration earlier.

The following diffs will slowly remove the namespace elaboration from the codegen pass. The reason why it is split up like this is because there are situations where moving the namespace elaboration early causes some problems.

The main effects that we need to address when we perform namespace elaboration early.
1.) There are places in codegen that expect (and match or compare against) non elaborated strings
To address this, we either special case the endo pass to not elaborate the names or we go through the codegen pass and fixup the namespacing by appending the `\` or `\HH\` to the strings.

# Backwards compatibility breaking changes:

HackC will no longer treat Hack collections as case insensitive. `map` will generate different bytecode than `Map`

HackC will no longer treat typehints in `is/as` statements as case insensitive. `$x is Nothing` will generate different bytecode than `$x is nothing`. Hack typechecker already makes this distinction.

Reviewed By: kmeht

Differential Revision: D17968303

fbshipit-source-id: 7d476f72181c9f53101ed92115e90b13d20a12e6