Simplify interface to HHBC to avoid nasty FFIs
Summary:
`hh_single_compile` binary calls one of the following
 - Emit_program.from_ast
 - Emit_program.fatal_program

followed by recording a time stamp, then it immediately calls
`Hhbc_Hhas.to_segments` to convert HHAS to a list of strings.

To avoid performance bottleneck of two-way Rust-Ocaml interop
and maintenance burden of HHAS FFI (keeping OCaml and Rust in Sync),
expose a higher-level interface that returns `string list` (segments):

- `Compile.from_ast`
- `Compile.fatal`

For now, TAST passed through the interface boundary (see `compile.mli`)
implying it'd need to be serialized to JSON and back to Rust TAST.
However, the rest of `do_compile` logic (i.e., call to parsing)
can be moved once lowerer is fully ported to Rust, thus skipping
the round-trip of AST that is likely to be a performance hog.

Finally, this allows for much greater flexibility in porting as less
of the details of hhbc/ package is exposed, and is a step forward
towards elimination of global mutable state:
- `Hhbc_options.compiler_options`
- `Hhbc_options.set_compiler_options`

Reviewed By: shiqicao

Differential Revision: D18132512

fbshipit-source-id: 40c4c411050749b24e616e7bc7f20d165f61354d