kill hm_nullsafe field in Thas_member
Summary:
see the small comment in typing.ml for how this works.

Another big chunk of this diff is to describe the super type in the error message generated by simplify_subtype. Otherwise, we'd get things like `Expected ?[unresolved]`. See the added test for an example of the kind of description it generates.

Reviewed By: manzyuk

Differential Revision: D17809248

fbshipit-source-id: 05688ba060861b099b441fe52cef972156e763e1