Run simplify_subtype during transitive closure
Summary:
The `sub_type` function currently proceeds in three stages:
1. Use `simplify_subtype` to reduce a statement `t <: u` to a logical proposition involving conjunction, disjunction, and atomic subtype assertions.
2. Use `props_to_env` to convert top-level conjuncts of the form `#v <: t` or `t <: #v` into lower and upper bounds on type variables in `env.tvenv`. At the same time, apply transitive closure on bounds of the form `t <: #v <: u` by calling *back* into `sub_type env t u`.
3. For any remaining empty disjunction (i.e. False), apply the associated nullary function to generate an error.

Going forward, we want to deal better with disjunctions, and not just pick the first one that is satisfiable. But to do this we want to maintain everything in a functional style. In particular, we want to move away from the `Errors.try` API that "undoes" errors in the global list.

In the above process, the recursive call to `sub_type` stands out: it can generate errors eagerly, which must be hidden using `Errors.try`. Instead, we should call back into `simplify_subtype` to extend the logical proposition to account for consequences of transitive closure. This is what we now do in `add_tyvar_lower_bound_and_close` and `add_tyvar_upper_bound_and_close`, which is called from `props_to_env`. We avoid using `Errors.try` in the case for disjunction in `props_to_env`, instead analysing the proposition that comes back, skipping unsatisfiable disjuncts, and choosing the first dijsunct that is not unsatisfiable.

Reviewed By: CatherineGasnier

Differential Revision: D18244350

fbshipit-source-id: 95caa1b3fb27f332ed0bb86a4146fd73341d7e95