Improve TLS code-gen
Summary:
Make some improvements to TLS code-gen. Generating code to interact
with TLS is a bit buggy at the moment, because most vasm instructions
do not handle the segment prefix correctly. Everything compiles, but
it silently generates incorrect instructions. If you try to use lea
(which cannot use segment registers), you likewise get something very
wrong.

Go through all of the vasm instructions that vasm-x64 supports, and
make sure they honor the segment prefix. Add an assert when emitting a
lea that there is no segment prefix set.

Improve the code-gen for emitting a TLS address, using a sub (where
appropriate) rather than an add of a large immediate. Add an offset
parameter to emitTLSLea(), which is useful if you're going to add an
offset anyways to the returned Vreg (it can be more efficient to fold
it into the lea calculation).

Reviewed By: markw65

Differential Revision: D18248016

fbshipit-source-id: 17243030bb95276fc5f14e35f07dd6115d8afe84