fix up darray/varray support in php serialize
Summary:
1. Expose option to preserve varrays through serialize_with_options
2. Teach unserialize() to deserialize 'y' into varray and 'Y' into darray by default. It still respects both m_forceArrays and RuntimeOption::EvalHackArrDVArrs flags.

'y' and 'Y' are HHVM-specific and are only understood by HHVM's VariableSerializer (which powers unserialize()).
This diff makes unserialize() produce varray for 'y' and darray for 'Y' by default. There is still force_darray option, which does as it says, and producing 'y' or 'Y' requires opt-in at serialization side. However we don't need explicit opt-in at deserialization time.

Reviewed By: ricklavoie

Differential Revision: D18264752

fbshipit-source-id: 387283626e82ae7d15707e607498da3ec2b7d3c8