Naming: SQLite connection should be held in a Mutex
Summary:
`rusqlite`'s Connection is not thread-safe. Therefore, to be used in a multithreaded context, it must be either held in a mutex, or a separate connection needs to be created in each thread.

Here, I'm taking the former approach. It's possible that we could later switch to the latter approach, or to using, e.g., channels to send names from the main thread to the child threads.

Reviewed By: Wilfred

Differential Revision: D18001692

fbshipit-source-id: b92e8050d51518ceb591e06a611f23128f752518