Leave stack and vmfp() in proper state when throwing from doFCall()
Summary:
Follow proper encapsulation rules and return the stack in consistent state
without the ActRec when throwing from doFCall().

We may call doFCall() from JIT, in which case the unwinder needs to know where
to continue unwinding. We need to make sure the catch of both the stub and the
caller execute exactly once and the best way to achieve that is to have a
consistent behavior.

Without this diff, D18095870 will call the caller's catch twice, using both
caller's and callee's frame. Once because of callee's saved RIP, the other time
because of stub's saved RIP.

Reviewed By: paulbiss

Differential Revision: D18238636

fbshipit-source-id: 40d3e4de09f8f17501825ee98bbf7b4bbd00559b