Add infrastructure to allow prologues to operate in stublogue mode
Summary:
Setup prologue in a stublogue context, then convert it to phplogue when
the frame gets written.

Keep track the duration of stublogue context, so that we can properly update
the native stack in JIT catch blocks.

The next diff will move most of the prologue logic to this section.

Reviewed By: ricklavoie

Differential Revision: D18054876

fbshipit-source-id: ec97917710695b3684b60684c7e63835cd0d19c6