Call all applicable catch traces when unwinding thru stublogues
Summary:
If we entered a C++ helper from a stublogue and that helper threw an exception,
we were skipping the catch trace of the caller of the stublogue. If there was
any stack spill across that call, we would fail to adjust the `rsp()`, crashing
few moments later.

The issue was that both the call to the stub and call from the stub operate
on the same logical VM frame, but the unwinder is capable of calling only one
catch block per VM frame, causing us to skip the catch block of the call into
the stub.

Fix this by adding `tc_unwind_resume_stublogue()` and `endCatchStublogueHelper`
stub that follows one step of the saved RIP chain without following RBP chain.

Reviewed By: markw65

Differential Revision: D18095870

fbshipit-source-id: 7c320ddf95dd8601009e235615591c83d6e7b00f