Remove aast macro
Summary:
This macro lets us avoid redundantly specifying the type arguments to the unnamed AST (`<Pos, (), (), ()>`), but we can do the same using type aliases (and already do so for the NAST).

This change fills out the `ast` module with type aliases and re-exports for the convenience of code dealing with unnamed ASTs, and removes the `aast!` macro in favor of qualified names in the `ast` module.

Reviewed By: Wilfred

Differential Revision: D18232163

fbshipit-source-id: 29f53c84bdcc2e150ad623cadfbfff9678745ed1