Fix width mismatch assertion caused by Bottoms
Summary:
When generating a Jmp to a DefLabel, and the input is a Bottom, we'll
treat it like a Bool. If the def SSATmp is not a Bool, we'll then emit
a zero extension instruction. This can cause a width mismatch because
we'll zero extend a 64-bit Vreg.

Fix this by not emitting the zero extension if the SSATmp is a
Bottom. Invent a new helper function to encapsulate this kind of
"subtype of, but not Bottom" check.

Differential Revision: D18304147

fbshipit-source-id: aed8cdbf8689a346b970616676d91d3b8753c54f