dummy get_gconst returns string
Summary:
This is just a tiny stepping stone that will be thrown away eventually. (washed away?)

The ocaml decl provider has only been hooked up to provide gconsts. Prior to this diff, the provider RPCd into the decl service and threw away the answer and returned an integer. Now with this diff it RPCs into the decl service and trusts that the decl service gave it a few raw utf8 bytes of string and returns that.

Neither is better than the other! But I wanted (1) to demonstrate that a payload placed in cachelib shared-mem by hh_decl could be successfully read from the ocaml side within shared-mem, (2) use something that will end up being represented as a block within cachelib, not just a primitive - because blocks will be more interesting.

Reviewed By: jewelpit

Differential Revision: D18277465

fbshipit-source-id: afd9174c2c0885a3d3e57d3ccd9633b000891fff