NakedOcamlPtr
Summary:
We already have ocamlrep for `UnsafeOcamlPtr`, which represents an unsafe pointer to an OCaml value (be it on the ocaml heap or not).

This diff introduces ocamlrep for `NakedOcamlPtr` which represents an pointer to anything, e.g. for when you want a pointer to some C struct, or to a shared memory base address.

I wasn't quite sure about naming conventions. The name "UnsafeOcamlPtr" sounds like it is an unsafe pointer to an ocaml value, in which case maybe this should be "UnsafeNakedPtr". But I wanted the word OCaml in there somewhere since this new type has no reason to exist other than for sake of OCaml.

Maybe I should have picked a name that doesn't mention "Naked" so that if we redo our ocaml stuff in future to have no naked pointers, then the current Rust pointer type can be changed? Nah. It's nicer to be explicit, and when that time comes then we can phase out this NakedOcamlPtr and replace it with a WrappedOcamlPtr.

Reviewed By: jewelpit

Differential Revision: D18248258

fbshipit-source-id: 7bd88b4465045bfd60daefb8caee60045afb06ed