naming 4: new API Naming_table.Consts.is_defined
Summary:
I'm trying to clean up some naming code to better understand (1) how+why it depends upon the naming table, (2) does it require positions.

Many uses of `Naming_table.Consts.get_pos` actually IGNORE the position; they use it only as a signal to see whether a const is defined or not. I'm adding a Naming_table API to indicate precisely that. This way, it'll be easier for us to identify those things that really do need pos, and figure out why.

(My conjecture is that pos is only needed for reporting text of error messages. Imagine if we removed pos from the naming-table, and anyone who wants pos must calculate it by parsing the file. If we only need pos for error messages, then this will be a net win).

Differential Revision: D17842069

fbshipit-source-id: 679c56e6cef9c67ed51c538e3c5863d7c8f84a7e