naming 6: new API Naming_table.*.get_fn
Summary:
I'm trying to clean up some naming code to better understand (1) how+why it depends upon the naming table, (2) does it require positions.

I've already covered many places where we used to call `Naming_table.XYZ.get_pos` solely for the sake of learning whether a name is defined.

This diff adds a new API `Naming_table.XYZ.get_fn` to look up the filename where something is defined.

If we discover that almost all uses merely need to know name and filename, and if we discover that the only scenarios that need actual line+column position information are error-reporting or other UI areas, then it'll be fine if our Naming Table saved-state simply omits positional information entirely. That would be more beautiful.

This diff is a step along the way to discovering whether we can do that.

Reviewed By: kmeht

Differential Revision: D17842226

fbshipit-source-id: e1f247b6a3a886da5ec2cfd5ff56cb99d1a2f95a