Fix misuse of Long_val
Summary:
`Val_long` takes a `long int` and returns an OCaml `value`. That's what this code intends to return.

```
/* Conversion macro names are always of the form  "to_from". */
/* Example: Val_long as in "Val from long" or "Val of long". */
#define Val_long(x)     ((intnat) (((uintnat)(x) << 1)) + 1)
#define Long_val(x)     ((x) >> 1)
```

If I understand correctly, this would basically return a value that is 1/4th of the actual value. Furthermore, the lowest bit wouldn't necessarily be 1. I'm guessing this is a bit used by the GC and if it's incorrectly set to 0 then a live value might get GC'd? Or something like that? I feel like I saw segfaults before this fix when using this function more widely but not afterwards.

Luckily we only use this function when we set `--sharedmemory-log-level 1`

Reviewed By: mroch

Differential Revision: D18203846

fbshipit-source-id: e47a14e6a60a9e7e7147af8bdf1fe760254036c7