naming 10: switch to Naming_table.Types.get_fn_and_kind
Summary:
I'm trying to clean up some naming code to better understand (1) how+why it depends upon the naming table, (2) does it require positions.

This diff adds a method `Naming_table.Types.get_fn_kind`, and sees how many places in the codebase need the kind and the filename (but don't care about the rest of the pos information).

The significance is (1) If the naming-table could get by with only storing kind and filename and not pos, then it'd be neater; (2) if we discover places which depend upon the kind but shouldn't, e.g. where the shape of the NAST depends upon global lookup of a symbol, then that smells pretty dodgy.

Reviewed By: kmeht

Differential Revision: D17840682

fbshipit-source-id: 5b9a882af7b9932c1cf7f55ce62dcb13351e7229