Consider call_user_func a built in and not a special function
Summary:
`call_user_func` is not a special function. It's a regular function that exists in the global namespace that we treat specially in naming.

This diff moves it out of the Special Function category. Behaviorally, this means that call_user_func can be namespaced and will work like any other function.

Reviewed By: shiqicao

Differential Revision: D18278243

fbshipit-source-id: 708890a0a2f89d47f6910ee5a1b83870a81bc742