Update the hover text for `__Soft` (#8581)
Summary:
The text _"will not be typechecked at runtime"_ suggests that HHVM is running hh_client (type checking) when running code. Not throwing a TypeError/Error more accurately describes the behavior of the runtime.

The text _"so callers will get checked"_ does not cover the case of a property, since they are assigned, not called.

Be double clear about continued execution after logging a warning.
Pull Request resolved: https://github.com/facebook/hhvm/pull/8581

Reviewed By: DavidSnider

Differential Revision: D18203994

fbshipit-source-id: f394d891e167a25f509c815541eec3dd01307e1e