Remove unused `file_edits` in `clientLsp`
Summary: We used to queue up the changes to files to be sent to `hh_server`. At some point, we changed it so that instead we keep track of the contents of open files, instead of the list of changes to those files. This means that we no longer have to keep track of those file edits. Remove that field from the environment, since it's not used anymore.

Differential Revision: D18311794

fbshipit-source-id: 0513dc1c77315b3c827a4a58dafee103e8bded39