TrustedAutoloadMap to read static SQLFacts DBs
Summary:
This diff allows the currently closed-source native autoloader to pull information from a read-only SQLite DB. We'll use this in situations where the underlying repository we're autoloading is also read-only, and where we can't necessarily crawl or watch this repository.

Most of the code in this commit is closed-source. I'll open source the native autoloader when it's stable.

Reviewed By: dneiter

Differential Revision: D18123293

fbshipit-source-id: 7c301af174f97cc806c3074b9181471157062d0b