Remove some uses of is_sub_type_LEGACY_DEPRECATED
Summary:
We would like to remove uses of the hacky `is_sub_type_LEGACY_DEPRECATED`. But in general, use of `is_sub_type` inside type *inference* is dodgy, because it returns `true` in cases of type variables. Instead, we should be generating constraints.

This is the first diff that reorganizes some of the code involved in Reactive features, so that instead of relying on predicates such as `is_sub_type`, it uses "assertions" that possibly generate constraints. More to come.

Reviewed By: manzyuk

Differential Revision: D18248024

fbshipit-source-id: 98d02b45e990e86831f403c3e3cebdb5905d5b98