Make HackC boolean options parseable from JSON bools
Summary:
Currently, `hh_single_compile` options are only parsed as bools
if they come as CLI arguments (flags) but not from JSON.

This is unfortunate because not only it is confusing;
JSON encoding should be lossless, yet the OCaml implementation
assumes that booleans are stored as "0" or "1", which are then coerced
to int then to bool.  Since Hhbc_options.ml provides no
canonical JSON representation of options, Rust port in `options.rs`
serialized those as JSON bools as opposed to lossily store them as strings.

This diff makes OCaml JSON deserialization more permissive, so that
it can correctly deserialize boolean options serialized by Rust.
A simpler alternative would be parse "true" or "false" strings
as ints 1 and 0, respectively (which would require no changes in Rust),
but this would could be more brittle/confusing.

Reviewed By: dabek

Differential Revision: D18353862

fbshipit-source-id: 9852275645e343fd888bd0a98b5b4702fca85890