Replace {Ld,St}Elem with more sensible opcodes
Summary:
`LdElem` and `StElem` are *weird* opcodes--they take a PtrToCell and
an Int and offset the pointer by that many **bytes** and derefence the result
as if it's a `TypedValue*`. This is weird for at least two reasons: it assumes
the PtrToCell is actually a pointer to an array of TypedValues and also
encourages all kinds of weird pointer arithmetic in the IR itself which is ....
odd.

We also don't want this kind of thing around anymore since we're trying to make
the layout of all kinds of aggregates more flexible--We shouldn't have an IR
opcode that lets anyone pretend they know what the underlying layout of a `Ptr`
into some aggregate is.

Since this was only ever used in a few places, just replace it, as follows:

For static properties: use LdMem since we can compute the destination pointer
statically and there's no need for the `Int` source.

For declPropInit lookups/writes: add `{Ld,St}ClsInitElem` opcodes to require
the offset to be computed statically.

For `Vector` lookups: just get the underlying `vec`--we don't need `LdElem` in
the first place!

For `Pair` lookups: add a `LdPairElem` opcode that coalesces the pointer
arithmetic and fetch into one op.

Reviewed By: ricklavoie

Differential Revision: D18185589

fbshipit-source-id: dcb4de5c22fdcd72bb8df643b0031548efcabc70