Memoize IdentifySymbolService TAST visitor
Summary:
When analyzing the performance of the LSP message "hover", I noticed that large files often took hundreds of milliseconds, even seconds, for each hover message to be processed.

Part of the problem was addressed by memoizing TAST computation.  The next step is memoizing the results of a TAST visitor in IdentifySymbolService that is used by both Highlight and Hover messages.

With this memoization, hover is now 50% faster.

Other changes included in this diff:
* Fixed a reference to Aast/Tast that made parsing the file slightly more awkward
* Added type annotations and names to parameters to the "go" function

Reviewed By: arxanas

Differential Revision: D18351680

fbshipit-source-id: 26e4aa3a137836068056f92e59a2b507d4b6f71a