Add source_text to entry
Summary:
In this diffstack, I aim to improve performance by reusing previous computations. Since pretty much every keystroke generates multiple messages, we should memoize the work of computing source_text for a file, with the aim of reusing this computed text rather than recomputing it.

This diff preserves the full fidelity source text that is already being computed.  Rather than throwing it away, we'll save it and reuse it for language server messages that need it (which will be enabled in a different diff in this stack)

Reviewed By: 2BitSalute

Differential Revision: D18379160

fbshipit-source-id: 72e44ac8cef473157c3f61fef11f530a67506134