Subsample perf warnings for SSATmp spanning across the calls
Summary:
This became super hot after mofarrell's diff that reuses the SSATmp for
`$this` or class ctx across the calls.

Let's subsample these perf warnings until ctx becomes a regular local and the
problem will go away (as we will be able to spill ctx into its actual slot
rather than to the native stack).

This also opens up an idea to split handling of values and types, so that we
can spill only the non-const value on the hot path and defer spilling the
statically known types to the cold paths.

Reviewed By: ricklavoie

Differential Revision: D18383131

fbshipit-source-id: a08379d639565c55f4a16d8b07d5cc81fc2e4502