Standardize behavior of autocomplete
Summary:
This diffstack memoizes certain computations to avoid redoing them on each language server message.

The "completion" message is a bit quirky.  During processing, it injects the text "AUTO332" into the file and uses this symbol as a way to identify which type of completion we want to use.  Because of this, the completion LSP message must do the following:
* Track the correct file contents in the real context, and
* Generate a fake "context" that has the AUTO332 edit.

This diff updates the behavior to work with the memoization functions.

Reviewed By: jewelpit

Differential Revision: D18383549

fbshipit-source-id: ae491e6f6912a3cd0236046d7eb52359523dc17c