Autoimported functions are not special functions
Summary:
There were some functions that were previously listed in the `SpecialFunctions` list when they were also in the list of autoimported functions.
- `fun`
- `class_meth`
- `meth_caller`
- `inst_meth`
- `invariant`

This diff removes those functions from the `SpecialFunctions` list and fixes up callsites in the compiler.

There should be no bytecode changes as a result of this diff, but it does help clean up some of the code in naming and namespace elaboration.

Reviewed By: Wilfred

Differential Revision: D18284042

fbshipit-source-id: 5c2d5c17597b4780f78c0ab8dff27e35f3337979