Make PseudoConstants case sensitive
Summary: Currently, pseudo / special consts like `__FUNCTION__` are case insensitive in the compiler. So `__fUnCtIon__` is treated the same as `__FUNCTION__`. There's no reason that they should be case sensitive - but there's no reason that they should be case insensitive. So go ahead and change it to be case sensitive. This is a backwards compatibility breaking change.

Reviewed By: kmeht

Differential Revision: D18308512

fbshipit-source-id: 569717ee149f4887dba08779b8e8ac5e6a12f720