Collect common iter args in a struct
Summary:
Some iterator ops have 5 or 6 immediates; no other opcodes have more than 4. That's getting to be unwieldy. I'd like to add a flags immediate so we can do further refcount analysis in HHBBC, but want to refactor these bytecodes before doing so.

Iterator ops also have very similar immediates. The only difference between IterInit and IterInitK is the key local; the only difference between IterInit and LIterInit is the base local. In this diff, we extract a set of four common immediates into an "IterArgs" struct: `iterId`, `keyId`, `valId`, and a `flags` immediate that currently only contains "baseConst", but which I'll extend soon.  For the value-only iterators, we allow `keyId` to be kNoKey, so we no longer need the "K" versions of these bytecodes. A lot of code can be de-duped using this representation; in particular, in particlar, bytecode.cpp, irgen-iter.cpp, and interp.cpp are all simpler.

A few notes on representation:
- The HHAS representation of IterArgs includes "NK" / "K:<keyLocal>" markers for the key and "V:<valLocal>" markers for the value so it's easy to distinguish them. (hackc actually flipped them twice, so that it was correct overall - see `instruction_sequence.ml`)
- The HHBC representation uses more bytes (e.g. we always include a key local). I don't think it will matter, but if we care to optimize we can just change `encodeIterArgs` / `decodeIterArgs`. I did ensure that value-only iters use 1 byte for keys.
- The `flags` immediate is specific to HHVM / HHBBC; it has no HHAS representation. This style has precedence with several flags in FCallArgs. I think it's even cleaner given that the entire flags struct is unused.

I'm going to simplify things further after this change. In particular, I will:
- Eliminate the "IterTypeOp" enum; we can just pass the flags + the local bit instead.
- Replace "IterData" with "IterArgs"; they're basically the same.
- Delete FIVE and SIX macros (assuming we don't want to have so many immediates anymore).

Since these further changes are all within HHVM, I'd rather not make them in this big boiler-plate-y diff/

After these changes, I'll get to the actual refcount optimization, which involves adding "KeyConst" and "ValConst" flags to IterArgs::Flags. These flags can be set for non-local iterators, which is why we want flags for them in the first place. "KeyConst" means "the key is not modified through its local". "ValConst" is similar.

Reviewed By: arnabde03

Differential Revision: D18362188

fbshipit-source-id: 73b6021ba686f2fe39a190e22e7c98efbb283568